refactor : Inputview 검증기능 분리
