+++ src/main/java/lotto/PrizeService.java
@@ -1,0 +1,4 @@
+package lotto;
+
+public class PrizeService {
+}
